chore : webRTC 컨테이너에 spaceId props 전달 버그 수정, (#281)
스냅샷 -> 작품 수정,
멤버 추가 모달 사용자 반응 추가
